+++ data-raw/make_data.R
@@ -5,6 +5,7 @@
 
 vdem    <- utils::read.csv("data-raw/eiv_vdem_panel.csv",  stringsAsFactors = FALSE)
 psid    <- utils::read.csv("data-raw/psid_wages_panel.csv", stringsAsFactors = FALSE)
+twins   <- utils::read.csv("data-raw/twins.csv", stringsAsFactors = FALSE)
 castle  <- utils::read.csv("data-raw/castle_panel.csv",    stringsAsFactors = FALSE)
 divorce <- utils::read.csv("data-raw/divorce_panel.csv",   stringsAsFactors = FALSE)
 fscore  <- utils::read.csv("data-raw/f_score_panel.csv",   stringsAsFactors = FALSE)
@@ -12,6 +13,7 @@ fscore  <- utils::read.csv("data-raw/f_score_panel.csv",   stringsAsFactors = FA
 if (!dir.exists("data")) dir.create("data")
 save(vdem,    file = "data/vdem.rda",    compress = "xz")
 save(psid,    file = "data/psid.rda",    compress = "xz")
+save(twins,   file = "data/twins.rda",   compress = "xz")
 save(castle,  file = "data/castle.rda",  compress = "xz")
 save(divorce, file = "data/divorce.rda", compress = "xz")
 save(fscore,  file = "data/fscore.rda",  compress = "xz")